requirements.txt
# Core dependencies
hvac>=1.2.1        # HashiCorp Vault client
pyyaml>=6.0.1      # YAML configuration file handling
urllib3>=2.0.7     # HTTP client (required by hvac)

# Type hints support
typing-extensions>=4.8.0

# Optional dependencies
requests>=2.31.0   # Required by hvac for HTTP requests
certifi>=2023.7.22 # Required for SSL certificate verification 